requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
frontend
tools
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence-transformers
langchain_huggingface
faiss-cpu
beautifulsoup4
streamlit
langchain_groq
fastapi
uvicorn
langserve[all]
langchain-chroma 
sqlalchemy
duckduckgo-search
sqlite